requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
gradio>=4.0.0
numpy>=1.24.0
requests>=2.31.0
